+++ marginal_approximation.h
@@ -5,6 +5,6 @@
 #ifndef PASTML_MARGINAL_APPROXI_H
 #define PASTML_MARGINAL_APPROXI_H
 
-void choose_likely_states(Tree *tree, int n);
+void choose_likely_states(Tree *tree, size_t n);
 
 #endif //PASTML_MARGINAL_APPROXI_H
